(FM-8273) Enable parallel acceptance from Windows
Prior to this commit running `litmus:acceptance:parallel` from a windows
node failed during the call to run the tests per target host. This was
due to the command string prepending the target host as an environment
variable before the execution, which does not work on Windows.

This commit updates the payloads creation to standardize the exec call
and enable passing an options hash, including environment variables to
send to the Open3 call.

This enables execution from a Windows node.

Co-authored-by: Glenn Sarti <glenn.sarti@outlook.com>