(FM-8296) Ensure serverspec helpers emit correctly
Prior to this commit the `run_bolt_task` and `bolt_upload_file` methods
did not correctly emit stdout and stderr - `run_bolt_task` also swallowed
the exit codes of a run.

This commit updates both methods to correctly capture and emit the stdout
and stderr streams and ensures that a specific exit code for a task is
captured instead of being overwritten to `255`.

The reason for these updates is that the `run_task` and `upload_file`
methods leveraged by `run_bolt_task` and `bolt_upload_file` respectively
do not return the same structured results as `run_command` and
`run script` - namely, it does not return the `exit_code`, `stderr`, and
`stdout` key-value pairs in the `result` hash.

Instead the output, exit code, and error must be extricated from the
object to be passed along.