(MAINT) Fix undeclared dependencies
The gemspec does not mention that this gem requires `parallel` and
`rspec`. Most users don't notice because those gems are already brought
in by other dependencies in the typical module's Gemfile.

The issue is when someone wants to leverage Litmus, but they aren't
doing modulet testing and want to avoid installing those other module
testing gems when they aren't needed, then the Litmus gem will fail to
install if `parallel` and `rspec` are not specifically added to the
Gemfile.

This change declares those two dependencies explicitly in the gemspec to
avoid this issue.