(MODULES-10018) Contain BoltSpec::Run methods
Prior to this commit the BoltSpec::Run methods were included at the
top level of the PuppetLitmus module namespace. This caused an
issue when leveraging the command() idiom in serverspec where it
attempted to resolve run_command() via the BoltSpec::Run method instead
of the appropriate serverspec method.

This commit contains the BoltSpec::Run methods into a class and adds
a shorthand referance to them for use in the rest of the PuppetLitmus
namespace.  This resolves the conflicts which prevented users from
leveraging idiomatic serverspec to test their modules and systems.