(MAINT) Add interpolate_powershell helper method
Prior to this commit running PowerShell commands against target hosts during an
acceptance test run required that the author both:

1. Wrap any PowerShell commands to ensure that they would be executed both
   locally and remotely in PowerShell, as bolt localhost execution does not
   run in PowerShell by default, and
2. Manually verify the interpolation of the code to be executed, accounting
   for passing from ruby to bolt to powershell/cmd and possibly beyond.

This commit adds a helper method to be available during acceptance test runs
which allows an author to pass a string for the scriptblock they want to run
in PowerShell; whatever that string evaluates to will be executed.

It does this first by base-64 encoding the command, enabling the use of the
EncodedCommand flag when invoking PowerShell.exe, which it then proceeds to do.