(GH-246) Fix inistall_modules_from_directory logic
Prior to this commit the install_modules_from_directory method
failed on Windows nodes when calling File.new on a folder; this
commit modifies the logic to instead grab the folder handle by
Dir.open and use that instead.

The logic in the same method for deleting the build folder
erroneously specified the builder instance which, at this point
in the execution, has not been instantiated yet. This commit
also modifies that logic to delete the target directory, which
is where the built module tars are placed.

These fixes support but do not resolve #246 filed by @findmyname666,
as the underlying problem around connections is caused by the
removal of the version from the inventory_hash in BoltSpec::Run.
For more information, see puppetlabs/bolt#1614.