Add run_bolt_plan helper
This change adds the ability for Litmus to run plans.

Plans are different from tasks in two major ways:

1. The target to run a plan against is a parameter to the plan, not to
Bolt. This means that target host is going to be a parameter to the plan
that the plan author has named whatever they like. As such the method
signature for the underlying `run_plan` helper in Boltspec has no target
param. As such
when calling a plan using this helper, it is up to the caller to know
the name of the parameter in the plan that will accept the target URI
and include the parameter and it's value in the `params` hash.

2. A plan is not a single executable that is copied to a target host
and run. As such there are no standard `exit_code`, `stdout` or `stderr`
nodes in the return we get from the Boltspec helper. We get only
`status` and `value` (A string array) if success, or `status` and
`value` (A complex object) if failure.

For backward compatibility and for ease of testing this new helper
creates a fake exit code, and it creates `stdout` and `stderr` keys in
the return object. The exit code is simply `0` if the plan succeeded and
`1` if it failed. The `stdout` key a string which is the `value` key of
the real result joined by line breaks. `Stderr` is is simply the `msg`
property of the `value` key if the result was a failure. If the result
was a success then `stderr` is nil.

Since `target_node_name` is not passed as a parameter to the underlying
Boltspec helper, it's only purpose here, and the reason it still exists
in this new function, is solely as a check to ensure that the target you
are about to pass to the plan exists in the inventory file you have
provided.