Convert inventory vars to JSON
Prior to this commit, the inventory_vars were passed to the provisioners
as a struct. The provisioners expect `vars` as a JSON formatted string,
which would cause failures in provisioning when `vars` was populated.
This commit changes the format of vars to be a string before it is
passed into `provision`.