(ISSUE 418) Accept a Target in Litmus Functions
This change will allow users to specify a target for each invocation of
a Litmus Puppet Helper. Specifying a target will not affect the current
value of the `ENV['TARGET_HOST']` variable.

Targets can be specified by using strings or symbols. Targets can be:
1. The name of a role assigned to a target via the `vars => roles` hash
of the target.
2. The name of a group in the inventory
3. The name of a specific target URI

The roles key in the inventory can be
1. Called either `Role` or `Roles`
2. Can have a single role `role: webserver` or an array:
  `roles:
    - IIS
    - Webserver
    - Frontend
  `
3. Values are searched in a case insensitive manner such that
  `'webserver' == 'Webserver'`

Closes #418

Co-authored-by: Daniel Carabas <daniel.carabas@puppet.com>