(CONT-193) Always build parallel task
Prior to this commit `litmus::acceptance` would check for the existance of
litmus_inventory before building the `parallel` task.

If the file is not present then the task would not be build and not
available in the list of tasks (`bundle exec rake --tasks`). It appears
that tasks are only evaluated at Runtime.

This causes an issue when calling tasks sequentially inside a task via
the api. In this particular case, the litmus_inventory file is only
present after the provision task has executed. This means that the task
list would have already been built without the acceptance task.

This would result in the following error:

`Don't know how to build task 'litmus:acceptance:parallel' (see
--tasks)`

This commit removes the check so that the task can always be built.
Checking for the existance of the `litmus_inventory` file is now scoped
around where we dynamically build tasks based on the nodes listed in the
inventory file.