(maint) - Update windows 2019 image
Prior to this commit, litmus would provision a windows-2019-core image
when performing testing against windows 2019 machines. This caused
failures on and blocks testing on the sqlserver module, with this error
"The following error occurred:
       You have selected a feature that is not supported on Windows Server Core."

The same tests pass on the other non-core windows machines.

This commit aims to update the windows-2019 image used, to a different
windows 2019 image in GCP.