feat: install an offline module using the tar.gz
What

This pull request allows the test of a module that have not been
released to the forge yet or simply needs to remain private. Please see #503

How

The module is built , then uploaded and the installation is done without
its dependencies.
dependencies are then parsed using the metadata.json file and installed
automatically

Signed-off-by: Julien Godin <julien.godin@camptocamp.com>