(CAT-1265) - Fix agent install on vagrant boxes
This commit fixes an issue where the puppet agent binary was not present
on the secure path of the vagrant boxes, so subsequent puppet operations
would fail.

It utilises the already present provision::fix_secure_path task in order
to make the puppet agent binary accessible to all sudo users on the box,
in most cases vagrant.