(maint) - Fix failing tests exiting 0
Prior to this commit, the spec tests would falsely exit 0, meaning that
the CI pipelines would pass and that failures would slip through the
cracks.

This was because we invoke rake tasks in some of our unit testing, which
explicity exit with 0 (in the cause the offending tasks was
litmus:check_connectivity). This exit code would interfere with the exit
code 1 of the failing rspec tests, leading to a false passing CI pipeline.

This now stubs the exit method, which is called during these rake tasks,
and ensures that it is not actually invoked and interferes with the
rspec exit code.